Change age type to unsigned int long

+++ mod3/strings.c
@@ -24,19 +24,19 @@ int main()
     // A string is a list/collection of characters (array)
     // Strings use %s as the placeholder
     char fName[SIZE]; 
-    int age;
+    unsigned int long  age;   // For unsigned use the %du, %lu, %llu
     char lName[SIZE]; 
     printf("Please enter you fName (< %d characters, NO space): ", SIZE);
    // When using arrays, DO NOT include the '&' operator in  your scanf
     scanf("%s", fName);
 
     printf("Enter your age: ");
-    scanf("%d", &age);
+    scanf("%lu", &age);
 
     printf("Please enter you lName (< %d characters, NO space): ", SIZE);
     scanf("%s", lName);
     
-    printf("Hi [%s][%s] [%d] years old\n", fName, lName, age);
+    printf("Hi [%s][%s] [%lu] years old\n", fName, lName, age);
 
     return 0;
 }
